refactor: move surah image map
