+++ data/surahImageMap.ts
