+++ jest.config.js
@@ -13,6 +13,7 @@ const customJestConfig = {
     '^@/infrastructure/(.*)$': '<rootDir>/src/infrastructure/$1',
     '^@/presentation/(.*)$': '<rootDir>/src/presentation/$1',
     '^@/lib/(.*)$': '<rootDir>/lib/$1',
+    '^@/data/(.*)$': '<rootDir>/data/$1',
     '^@/types/(.*)$': '<rootDir>/types/$1',
     '^@/(.*)$': '<rootDir>/src/$1',
   },
