+++ lib/api/__tests__/chapters.test.ts
@@ -2,7 +2,7 @@ import { getChapters, getSurahCoverUrl, getSurahList } from '@/lib/api/chapters'
 import { API_BASE_URL } from '@/lib/api';
 import { Chapter, Surah } from '@/types';
 
-jest.mock('@/presentation/(features)/surah/lib/surahImageMap', () => ({
+jest.mock('@/data/surahImageMap', () => ({
   surahImageMap: { 1: 'test.jpg' },
 }));
